requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
googletrans==4.0.0rc1
# Editable install with no version control (gpt==0.0.0)
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
openai==0.6.3
requests==2.25.1
rfc3986==1.4.0
sniffio==1.2.0
tqdm==4.60.0
urllib3==1.26.4certifi==2020.12.5
chardet==3.0.4
click==7.1.2
googletrans==4.0.0rc1
# Editable install with no version control (gpt==0.0.0)
-e /Users/jun/gpt
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
openai==0.6.3
python-dotenv==0.17.0
requests==2.25.1
rfc3986==1.4.0
sniffio==1.2.0
tqdm==4.60.0
urllib3==1.26.4
